Added some wrappers for the Windows platform and updated various pieces of code to use them.
Implemented a function for asynchronously monitoring a spawned process to misc.py.
This made the ioctl's and select on stdin within lib/ansible/plugins/connection/local.py
unnecessary and allowed the loop to be replaced with a closure.

Implemented an object that abstracts away the ioctl-style file-locking in exchange
for an object-based contextable style of locking. One was implemented for Windows
and the older ioctl-based style was left for posix.

Added a portable way of determining the Console's dimensions using the ctypes
module. The older ioctl-based one was also used, but utilizing a ctypes structure
instead of using struct.unpack to snag it's values.