Improved support for using generators to the utils.misc.process class. This allows one to deal with i/o a lot simplier than doing a for-loop over select. Also added support for capturing exceptions in a sub-thread in order to dispatch them to the main Python thread.
Forgot to include the full module path when importing the windows wrapper
modules around pwd and grp. Fixed them.

Removed the unncessary reference to the TIOCGWINSZ constant since that's now
handled within the utils.misc module.

process.wait() was deadlocking if the process had already died due to incorrectly checking the result of program.poll()
sub-threads were not getting deleted properly when the process was stopped due to indenting.
Added a 'paused' option for specifying whether to start the process in a paused state.